O365 New User Provisioning
This file has been sanitized, there are some hard coded entries that need to be replaced based on your organization's environment. The entries that need to be replaced are INALLCAPS, with exception to the O365 license lines at the end, those do not need to be edited.

This script was written for an AD structured with user accounts in an OU "/User Accounts/CityName/Department format/" but can easily be modified for other structures (modify the New-RemoteMailbox line).

The organization utilized Exclaimer Cloud signature product, the $Exclaimer lines relate to the AD group for each department's signatures. Remove as needed.